Use prop "helmetKey" if it's present as primary attribute.
This allows more fine-grained control over which tags are duplicated vs
overwritten.
Tags of the same type with the same value in "helmetKey"
will be overwritten, even if the would normally be duplicated.
Tags of the same type with different values in "helmetKey" will be
duplicated, even if they would normally be overwritten.